functions/query_handler/requirements.txt
functions-framework
google-cloud-aiplatform
vertexai
google-api-python-client
google-auth-oauthlib
python-docx
PyMuPDF
